# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Express — Hợp đồng tích hợp backend",
    "version": "0.2.0",
    "description": "Frontend HTTP-only ngày 02/10/2026. Các endpoint cũ giữ nguyên. Register tạo shop rồi yêu cầu đăng nhập; không tự tạo session. Cookie và CSRF cần backend triển khai."
  },
  "servers": [
    {
      "url": "https://localhost:7001/api/v1",
      "description": "Ví dụ; thay bằng URL backend thực tế"
    }
  ],
  "security": [
    {
      "cookieSession": []
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Đăng nhập và thiết lập cookie phiên; không trả role do client tự chọn.",
        "description": "Quyền đề xuất: Ẩn danh. Hợp đồng tích hợp suy ra từ frontend, chưa xác nhận backend đã triển khai.",
        "tags": [
          "auth"
        ],
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Thành công; trả JSON trực tiếp theo schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionUser"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập/hết phiên",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy trong phạm vi được phép",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "409": {
            "description": "Xung đột trạng thái, báo giá, năng lực hoặc kỳ đối soát",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi máy chủ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              },
              "example": {
                "email": "{{email}}",
                "password": "{{password}}"
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "Hủy phiên và trả JSON {success:true}; frontend chưa xử lý 204.",
        "description": "Quyền đề xuất: Đã đăng nhập. Hợp đồng tích hợp suy ra từ frontend, chưa xác nhận backend đã triển khai.",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "cookieSession": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Thành công; trả JSON trực tiếp theo schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResult"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập/hết phiên",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy trong phạm vi được phép",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "409": {
            "description": "Xung đột trạng thái, báo giá, năng lực hoặc kỳ đối soát",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi máy chủ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/register": {
      "post": {
        "operationId": "register",
        "summary": "Đăng ký tài khoản shop",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo thành công; frontend quay lại đăng nhập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResult"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "409": {
            "description": "Email đã tồn tại"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "UserRole": {
        "type": "string",
        "enum": [
          "shop",
          "staff",
          "operations"
        ]
      },
      "SessionUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "required": [
          "id",
          "name",
          "role"
        ]
      },
...
      "LoginInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LogoutResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "RegisterInput": {
        "type": "object",
        "required": [
          "fullName",
          "shopName",
          "phone",
          "email",
          "password"
        ],
        "properties": {
          "fullName": {
            "type": "string",
            "maxLength": 100
          },
          "shopName": {
            "type": "string",
            "maxLength": 150
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "writeOnly": true
          }
        }
      },
      "RegisterResult": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "Express.Session",
        "description": "Tên cookie đề xuất; dùng HttpOnly server session. Không phải JWT hiện có."
      }
    }
  }
}